fix some errors and remove unnecessary envs from quiz_service

+++ srcs/backend/main_service/src/services/fileService.js
@@ -26,8 +26,11 @@ export const saveAvatar = async (userId, file) => {
             throw new HttpException(400, 'inalid avatar image');
         }
     } catch (err) {
-        await fs.unlink(physicalPath).catch(() => {});
-        throw err;
+        if (result.class === 'non valid profile')
+        {
+            await fs.unlink(physicalPath).catch(() => {});
+            throw err;
+        }
     }
     return {
         type :'avatar',
